add mindfulness page
add mindfulness page; correct the mixed use between mindfulness and affirmation before; add a simple test for created page
