+++ __tests__/affirmation.test.tsx
@@ -1,0 +1,16 @@
+import '@testing-library/jest-dom'
+import { render, screen } from '@testing-library/react'
+import Page from '../app/affirmation/page'
+
+jest.mock('next/navigation', () => jest.requireActual('next-router-mock'))
+jest.mock('../lib/zoomapi', () => jest.requireActual('../lib/fakezoomapi'));
+
+describe('Page', () => {
+  it('renders a heading', () => {
+    render(<Page />)
+
+    const heading = screen.getByRole('heading', { level: 1 })
+
+    expect(heading).toBeInTheDocument()
+  })
+})
